requirements.txt
anthropic
feedparser
requests
python-dotenv
openai
google-genai
Pillow
certifi
openpyxl
